Catch assertion error from gitHub-flask library
The GitHub-flask library will raise an assertion error if GitHub returns
a status code that isn't 4xx or 2xx. This caused crashes when GitHub
returned a 503 error.

For now I'm just catching the error and reporting it. In the future I'll
probably submit a pull request to the library that makes it do something
more sensible.