Redirect to index with params on error
When an error occurs, the user is now redirected to the main page with
the data the user already entered encoded in the URL. This data is then
automatically filled into the fields.

This allows the user to fix their existing request instead of having to
re-enter everything.

Fixes issue #2.